setup unittest for jwt decoder
